Changed data type of session into long

+++ Sources/Classes/Public/RSConfig.h
@@ -19,7 +19,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic) int sleepTimeout;
 @property (nonatomic) int logLevel;
 @property (nonatomic) int configRefreshInterval;
-@property (nonatomic) int sessionInActivityTimeOut;
+@property (nonatomic) long sessionInActivityTimeOut;
 @property (nonatomic) bool trackLifecycleEvents;
 @property (nonatomic) bool recordScreenViews;
 @property (nonatomic) bool enableBackgroundMode;
